Rename xwalk.scene to xwalk.animation.
